fixed bots getting stuck; longer range detection
if 2 bots were traveling on the same line between two nodes in op directions they would get stuck on each other; if bots do not reach their destination within 1 second - recalculate the path

+++ code/game/ai_main.h
@@ -234,7 +234,7 @@ typedef struct bot_state_s
 
 	float				destinationGrabTime;
 	float				wpSeenTime;
-	float				wpTravelTime;
+	int					wpTravelTime;
 	float				wpDestSwitchTime;
 	float				wpSwitchTime;
 	float				wpDestIgnoreTime;
